[CONV] Remove convmerge layer
There are 3 psum buffers for the execution on intermittent power, and
the output activation will only be placed in one of them.
However, the next layer can not detect which input buffer be used to
place real input. Therefore, adjust the psum buffer version at begin to
make the output activation will be placed in buffer 0. The next layer
can directly fetch real input from buffer 0.